# 参考及文档说明/接口文档/智能工单系统 API.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "智能工单系统 API",
    "description": "智能工单管理系统接口文档",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:10001",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/workorder": {
      "post": {
        "tags": [
          "工单管理"
        ],
        "summary": "创建工单",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkOrderCreateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLong"
                }
              }
            }
          }
        }
      }
    },
    "/workorder/{id}/withdraw": {
      "post": {
        "tags": [
          "工单管理"
        ],
        "summary": "撤回/取消工单",
        "operationId": "withdraw",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/workorder/{id}/review": {
      "post": {
        "tags": [
          "工单管理"
        ],
        "summary": "审核工单",
        "operationId": "review",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/workorder/{id}/resubmit": {
      "post": {
        "tags": [
          "工单管理"
        ],
        "summary": "重新提交工单",
        "operationId": "resubmit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkOrderCreateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/workorder/{id}/process": {
      "post": {
        "tags": [
          "工单管理"
        ],
        "summary": "处理完成",
        "operationId": "process",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/workorder/{id}/dispatch": {
      "post": {
        "tags": [
          "工单管理"
        ],
        "summary": "派单/转派",
        "operationId": "dispatch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispatchDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/workorder/{id}/accept": {
      "post": {
        "tags": [
          "工单管理"
        ],
        "summary": "验收工单",
        "operationId": "accept",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/workorder/{id}": {
      "get": {
        "tags": [
          "工单管理"
        ],
        "summary": "工单详情",
        "operationId": "detail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultWorkOrderDetailVO"
                }
              }
            }
          }
        }
      }
    },
    "/workorder/stats": {
      "get": {
        "tags": [
          "工单管理"
        ],
        "summary": "工单统计",
        "operationId": "stats",
        "parameters": [
          {
            "name": "orderType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultWorkOrderStatsVO"
                }
              }
            }
          }
        }
      }
    },
    "/workorder/page": {
      "get": {
        "tags": [
          "工单管理"
        ],
        "summary": "工单分页列表",
        "operationId": "page",
        "parameters": [
          {
            "name": "current",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 10
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "orderType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageResultWorkOrderVO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResultVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {},
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "WorkOrderCreateDTO": {
        "type": "object",
        "description": "创建工单请求参数",
        "properties": {
          "title": {
            "type": "string",
            "description": "工单标题",
            "maxLength": 200,
            "minLength": 0
          },
          "content": {
            "type": "string",
            "description": "工单描述,可空",
            "maxLength": 5000,
            "minLength": 0
          },
          "orderType": {
            "type": "integer",
            "format": "int32",
            "description": "工单类型:1故障报修 2资源申请 3需求变更",
            "maximum": 3,
            "minimum": 1
          },
          "priority": {
            "type": "integer",
            "format": "int32",
            "description": "优先级:1高 2中 3低",
            "maximum": 3,
            "minimum": 1
          },
          "expireTime": {
            "type": "string",
            "format": "date-time",
            "description": "超时时间,可空;不传则默认 72 小时后到期"
          },
          "resources": {
            "type": "array",
            "description": "资源明细列表,可空",
            "items": {
              "$ref": "#/components/schemas/WorkOrderResourceDTO"
            }
          }
        },
        "required": [
          "orderType",
          "priority",
          "title"
        ]
      },
      "WorkOrderResourceDTO": {
        "type": "object",
        "description": "工单资源明细",
        "properties": {
          "resourceType": {
            "type": "string",
            "description": "资源类别:服务器/带宽/软件许可 等",
            "maxLength": 50,
            "minLength": 0
          },
          "resourceName": {
            "type": "string",
            "description": "资源名称/规格描述",
            "maxLength": 200,
            "minLength": 0
          },
          "quantity": {
            "type": "number",
            "description": "数量,必须大于0",
            "minimum": 0.01
          },
          "unit": {
            "type": "string",
            "description": "单位:台/MBps/个 等",
            "maxLength": 20,
            "minLength": 0
          },
          "remark": {
            "type": "string",
            "description": "申请说明,可空",
            "maxLength": 500,
            "minLength": 0
          }
        },
        "required": [
          "quantity",
          "resourceName",
          "resourceType"
        ]
      },
      "ResultLong": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "integer",
            "format": "int64"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "WithdrawDTO": {
        "type": "object",
        "description": "工单撤回/取消请求参数",
        "properties": {
          "remark": {
            "type": "string",
            "description": "撤回/取消原因,可空",
            "maxLength": 500,
            "minLength": 0
          }
        }
      },
      "ReviewDTO": {
        "type": "object",
        "description": "工单审核请求参数",
        "properties": {
          "approved": {
            "type": "boolean",
            "description": "是否通过:true 通过(→待派单) false 驳回(→已驳回)"
          },
          "remark": {
            "type": "string",
            "description": "审核意见/驳回原因,可空",
            "maxLength": 500,
            "minLength": 0
          }
        },
        "required": [
          "approved"
        ]
      },
      "ProcessDTO": {
        "type": "object",
        "description": "处理完成请求参数",
        "properties": {
          "remark": {
            "type": "string",
            "description": "处理结果说明,可空",
            "maxLength": 500,
            "minLength": 0
          }
        }
      },
      "DispatchDTO": {
        "type": "object",
        "description": "派单/转派请求参数",
        "properties": {
          "handlerId": {
            "type": "integer",
            "format": "int64",
            "description": "目标处理人用户ID"
          },
          "remark": {
            "type": "string",
            "description": "派单/转派说明,可空",
            "maxLength": 500,
            "minLength": 0
          }
        },
        "required": [
          "handlerId"
        ]
      },
      "AcceptDTO": {
        "type": "object",
        "description": "工单验收请求参数",
        "properties": {
          "approved": {
            "type": "boolean",
            "description": "是否通过:true 通过(→已完成) false 退回(→处理中)"
          },
          "remark": {
            "type": "string",
            "description": "验收意见/退回原因,可空",
            "maxLength": 500,
            "minLength": 0
          }
        },
        "required": [
          "approved"
        ]
      },
...
      "ResultWorkOrderDetailVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/WorkOrderDetailVO"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "WorkOrderDetailVO": {
        "type": "object",
        "description": "工单详情",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "工单ID"
          },
          "orderNo": {
            "type": "string",
            "description": "工单编号"
          },
          "userId": {
            "type": "integer",
            "format": "int64",
            "description": "提单人ID"
          },
          "departmentId": {
            "type": "integer",
            "format": "int64",
            "description": "归属部门ID"
          },
          "handlerId": {
            "type": "integer",
            "format": "int64",
            "description": "处理人ID,派单前为空"
          },
          "title": {
            "type": "string",
            "description": "工单标题"
          },
          "content": {
            "type": "string",
            "description": "工单描述"
          },
          "orderType": {
            "type": "integer",
            "format": "int32",
            "description": "工单类型:1故障报修 2资源申请 3需求变更"
          },
          "priority": {
            "type": "integer",
            "format": "int32",
            "description": "优先级:1高 2中 3低"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "状态:0待审核 1待派单 2处理中 3待验收 4已完成 5已驳回 6已取消 7已超时"
          },
          "statusDesc": {
            "type": "string",
            "description": "状态名称"
          },
          "remark": {
            "type": "string",
            "description": "备注/驳回原因/关闭说明"
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间"
          },
          "expireTime": {
            "type": "string",
            "format": "date-time",
            "description": "超时时间"
          },
          "resources": {
            "type": "array",
            "description": "资源明细",
            "items": {
              "$ref": "#/components/schemas/WorkOrderResource"
            }
          },
          "logs": {
            "type": "array",
            "description": "操作日志(按时间升序)",
            "items": {
              "$ref": "#/components/schemas/WorkOrderOperateLog"
            }
          }
        }
      },
      "WorkOrderOperateLog": {
        "type": "object",
        "description": "工单操作日志实体",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "主键"
          },
          "orderId": {
            "type": "integer",
            "format": "int64",
            "description": "工单ID"
          },
          "operatorId": {
            "type": "integer",
            "format": "int64",
            "description": "操作人ID(user.id)"
          },
          "operateType": {
            "type": "integer",
            "format": "int32",
            "description": "操作事件:1提交 2审核通过 3审核驳回 4派单 5处理完成 6验收通过 7退回处理 8转派 9撤回/取消 10超时关闭"
          },
          "fromStatus": {
            "type": "integer",
            "format": "int32",
            "description": "变更前状态(0-7),首次提交为 NULL"
          },
          "toStatus": {
            "type": "integer",
            "format": "int32",
            "description": "变更后状态(0-7)"
          },
          "toUserId": {
            "type": "integer",
            "format": "int64",
            "description": "派单/转派目标处理人ID,可空"
          },
          "remark": {
            "type": "string",
            "description": "操作备注,可空"
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "description": "操作时间"
          }
        }
      },
      "WorkOrderResource": {
        "type": "object",
        "description": "工单资源明细实体",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "主键"
          },
          "orderId": {
            "type": "integer",
            "format": "int64",
            "description": "所属工单ID"
          },
          "resourceType": {
            "type": "string",
            "description": "资源类别:服务器/带宽/软件许可 等"
          },
          "resourceName": {
            "type": "string",
            "description": "资源名称/规格描述"
          },
          "quantity": {
            "type": "number",
            "description": "数量"
          },
          "unit": {
            "type": "string",
            "description": "单位:台/MBps/个 等"
          },
          "remark": {
            "type": "string",
            "description": "申请说明,可空"
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          }
        }
      },
      "ResultWorkOrderStatsVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/WorkOrderStatsVO"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "StatusCount": {
        "type": "object",
        "description": "状态计数",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "状态码 0-7"
          },
          "statusDesc": {
            "type": "string",
            "description": "状态名称"
          },
          "count": {
            "type": "integer",
            "format": "int64",
            "description": "该状态下的工单数"
          }
        }
      },
      "WorkOrderStatsVO": {
        "type": "object",
        "description": "工单统计",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "当前可见范围内的工单总数"
          },
          "statusCounts": {
            "type": "array",
            "description": "各状态工单数,含数量为 0 的状态,按状态码升序",
            "items": {
              "$ref": "#/components/schemas/StatusCount"
            }
          }
        }
      },
      "PageResultWorkOrderVO": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkOrderVO"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "current": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ResultPageResultWorkOrderVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageResultWorkOrderVO"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "WorkOrderVO": {
        "type": "object",
        "description": "工单列表项",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "工单ID"
          },
          "orderNo": {
            "type": "string",
            "description": "工单编号"
          },
          "userId": {
            "type": "integer",
            "format": "int64",
            "description": "提单人ID"
          },
          "departmentId": {
            "type": "integer",
            "format": "int64",
            "description": "归属部门ID"
          },
          "handlerId": {
            "type": "integer",
            "format": "int64",
            "description": "处理人ID,派单前为空"
          },
          "title": {
            "type": "string",
            "description": "工单标题"
          },
          "orderType": {
            "type": "integer",
            "format": "int32",
            "description": "工单类型:1故障报修 2资源申请 3需求变更"
          },
          "priority": {
            "type": "integer",
            "format": "int32",
            "description": "优先级:1高 2中 3低"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "状态:0待审核 1待派单 2处理中 3待验收 4已完成 5已驳回 6已取消 7已超时"
          },
          "statusDesc": {
            "type": "string",
            "description": "状态名称"
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间"
          },
          "expireTime": {
            "type": "string",
            "format": "date-time",
            "description": "超时时间"
          }
        }
      },
...
    }
  }
}